refactor: 서비스 간 순환 참조 오류 해결
